fix: use float for dateFrom/dateTo to avoid 32-bit overflow for post-2038 timestamps
@Slot(int, int) and @Property(int) map to Qt's 32-bit C++ int. Unix timestamps
for years beyond 2038 (e.g. 2155) overflow and wrap to a value in 2018/2019,
causing the timeline histogram to select the wrong year on click.

Fix: change @Property(int) -> @Property(float) for dateFrom/dateTo, and
@Slot(int,int) -> @Slot(float,float) for setDateFilter (casting to int
internally). QML bindings changed from 'int' to 'double' for _dateFrom/_dateTo.

Triggered by an image with a corrupt EXIF DateTimeOriginal of 2155-10-23.

+++ src/exif_turbo/ui/view_models/app_controller.py
@@ -522,13 +522,13 @@ class AppController(QObject):
     def exiftoolVersion(self) -> str:
         return self._exiftool_version
 
-    @Property(int, notify=dateFilterChanged)
-    def dateFrom(self) -> int:
-        return self._date_from if self._date_from is not None else 0
+    @Property(float, notify=dateFilterChanged)
+    def dateFrom(self) -> float:
+        return float(self._date_from) if self._date_from is not None else 0.0
 
-    @Property(int, notify=dateFilterChanged)
-    def dateTo(self) -> int:
-        return self._date_to if self._date_to is not None else 0
+    @Property(float, notify=dateFilterChanged)
+    def dateTo(self) -> float:
+        return float(self._date_to) if self._date_to is not None else 0.0
 
     @Property(str, notify=yearCountsChanged)
     def yearCounts(self) -> str:
@@ -1255,11 +1255,11 @@ class AppController(QObject):
         self._run_search()
         return self._query_text
 
-    @Slot(int, int)
-    def setDateFilter(self, date_from: int, date_to: int) -> None:
+    @Slot(float, float)
+    def setDateFilter(self, date_from: float, date_to: float) -> None:
         """Set captured_at filter bounds (Unix timestamps, 0 = unset)."""
-        new_from = date_from if date_from != 0 else None
-        new_to = date_to if date_to != 0 else None
+        new_from = int(date_from) if date_from != 0 else None
+        new_to = int(date_to) if date_to != 0 else None
         if new_from == self._date_from and new_to == self._date_to:
             return
         self._date_from = new_from
